chore: break out body of LedgerOpenApp into a separate component
