+++ src/components/LedgerOpenApp/components/AssetOnLedger.tsx
